moving network connection interface into base, removing unusued clientConnection
